backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
django-environ==0.11.2
django-storages==1.14.2
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
